feat(130): add support for manifest-v3 CSP
The new function to compute defaults for the manifest file should be
able to maintain the old defaults for manifest-v2 exactly as they were,
while also allowing manifest-v3 compatible objects.

The function will also emit a warning when an unsupported manifest
version is encountered, but for the sake of future-proofing, no errors
are emitted, and the compiler process is not halted.

Signed-off-by: Buckley Ross <buckleyross42@gmail.com>
Change-Id: Id5b36209cb842b0c97698319574751c0d72b767e